chore: convert idempotency middleware to MiddlewareRenderable
